Move qgsoptions.cpp to new app/options subdirectory
